requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
adal==1.2.7
    # via msrestazure
azure-common==1.1.27
    # via
    #   azure-mgmt-cosmosdb
    #   azure-mgmt-iothub
    #   azure-mgmt-keyvault
    #   azure-mgmt-resource
    #   azure-mgmt-servicebus
    #   azure-mgmt-signalr
    #   azure-mgmt-storage
    #   azure-mgmt-web
azure-core==1.14.0
    # via
    #   -r requirements.in
    #   azure-cosmos
    #   azure-identity
    #   azure-mgmt-core
azure-cosmos==4.2.0
    # via -r requirements.in
azure-identity==1.6.0
    # via -r requirements.in
azure-iot-hub==2.4.0
    # via -r requirements.in
azure-mgmt-core==1.2.2
    # via
    #   azure-mgmt-cosmosdb
    #   azure-mgmt-iothub
    #   azure-mgmt-keyvault
    #   azure-mgmt-resource
    #   azure-mgmt-servicebus
    #   azure-mgmt-storage
    #   azure-mgmt-web
azure-mgmt-cosmosdb==6.3.0
    # via -r requirements.in
azure-mgmt-iothub==2.0.0
    # via -r requirements.in
azure-mgmt-keyvault==9.0.0
    # via -r requirements.in
azure-mgmt-resource==18.0.0
    # via -r requirements.in
azure-mgmt-servicebus==6.0.0
    # via -r requirements.in
azure-mgmt-signalr==0.4.0
    # via -r requirements.in
azure-mgmt-storage==18.0.0
    # via -r requirements.in
azure-mgmt-web==3.0.0
    # via -r requirements.in
azure.core==1.14.0
    # via azure-iot-hub
certifi==2020.12.5
    # via
    #   msrest
    #   requests
    #   uamqp
cffi==1.14.5
    # via cryptography
chardet==4.0.0
    # via requests
cryptography==3.4.7
    # via
    #   adal
    #   azure-identity
    #   msal
    #   pyjwt
gitdb==4.0.7
    # via gitpython
gitpython==3.1.17
    # via -r requirements.in
idna==2.10
    # via requests
isodate==0.6.0
    # via msrest
msal-extensions==0.3.0
    # via azure-identity
msal==1.12.0
    # via
    #   azure-identity
    #   msal-extensions
msrest==0.6.21
    # via
    #   -r requirements.in
    #   azure-iot-hub
    #   azure-mgmt-cosmosdb
    #   azure-mgmt-iothub
    #   azure-mgmt-keyvault
    #   azure-mgmt-resource
    #   azure-mgmt-servicebus
    #   azure-mgmt-signalr
    #   azure-mgmt-storage
    #   azure-mgmt-web
    #   msrestazure
msrestazure==0.6.4
    # via
    #   -r requirements.in
    #   azure-mgmt-signalr
oauthlib==3.1.0
    # via requests-oauthlib
portalocker==1.7.1
    # via msal-extensions
pycparser==2.20
    # via cffi
pyjwt[crypto]==2.1.0
    # via
    #   adal
    #   msal
python-dateutil==2.8.1
    # via adal
pywin32==300
    # via portalocker
requests-oauthlib==1.3.0
    # via msrest
requests==2.25.1
    # via
    #   adal
    #   azure-core
    #   azure.core
    #   msal
    #   msrest
    #   requests-oauthlib
six==1.16.0
    # via
    #   azure-core
    #   azure-cosmos
    #   azure-identity
    #   azure.core
    #   isodate
    #   msrestazure
    #   python-dateutil
    #   uamqp
smmap==4.0.0
    # via gitdb
typing-extensions==3.10.0.0
    # via gitpython
uamqp==1.4.0
    # via azure-iot-hub
urllib3==1.26.5
    # via requests
